kbfsgit: for fetches, use on-demand storer and skip compression
By default, go-git reads all data objects into memory while building a
packfile to push to a remote (which is the operation we use for
fetches).  If the repo is too big, this results in extraordinary
memory usage.

Instead, use a shim storer to prevent loading object data into memory
except when it is needed (via the `Reader()` object call, or to verify
the object type or size).

go-git uses delta compression to build the packfile, but it is very
slow.  It also messes up the on-demand storage plan, because go-git
modifies data in memory and expects it to survive all the way to
flushing it to the packfile.  If we read data on demand and then
forget about it, go-git will _think_ it compressed an object, but
those changes will get lost, and the resulting packfile will be
corrupt.  Plus, we don't even need delta compression, because we're
not transferring the objects over a network, only to a local git repo.
So we turn delta compression off entirely.

Issue: KBFS-2406